fixes for sp-only compilation

+++ sp/initBathymetry_formula.h
@@ -1,4 +1,4 @@
-inline void initBathymetry_formula(const float *coords, float *values, const double *time) {
+inline void initBathymetry_formula(const float *coords, float *values, const float *time) {
   float x = coords[0];
   float y = coords[1];
   float t = *time;
