+++ sp/initEta_formula.h
@@ -1,4 +1,4 @@
-inline void initEta_formula(const float *coords, float *values, const double *time) {
+inline void initEta_formula(const float *coords, float *values, const float *time) {
   float x = coords[0];
   float y = coords[1];
   float t = *time;
